Fix and move the reactor call decorator to util.py and refactor all the unit tests to use it.
