+++ tests/test_candidates.py
@@ -7,15 +7,17 @@
 # disable C0301, Line too long.  the auto generated tests do not conform to this rule.
 # pylint: disable=C0301
 
-from fractions import gcd
 from itertools import combinations, islice
 from time import time
 
 from ..candidate import CANDIDATE_ELIGIBLE_DELAY
 from ..logger import get_logger
 from ..tool.tracker import TrackerCommunity
+from ..util import call_on_reactor_thread
 from .debugcommunity.community import DebugCommunity
-from .dispersytestclass import DispersyTestFunc, call_on_mm_thread
+from .dispersytestclass import DispersyTestFunc
+
+
 logger = get_logger(__name__)
 
 
@@ -327,7 +329,7 @@ class TestCandidates(DispersyTestFunc):
 
         return [candidate for flags, candidate in zip(all_flags, candidates) if filter_func(flags, candidate)]
 
-    @call_on_mm_thread
+    @call_on_reactor_thread
     def check_candidates(self, all_flags):
         assert isinstance(all_flags, list)
         assert all(isinstance(flags, str) for flags in all_flags)
@@ -412,7 +414,7 @@ class TestCandidates(DispersyTestFunc):
             candidate = community.dispersy_get_walk_candidate()
             self.assertEquals(candidate, None)
 
-    @call_on_mm_thread
+    @call_on_reactor_thread
     def test_get_introduce_candidate(self, community_create_method=DebugCommunity.create_community):
         community = community_create_method(self._dispersy, self._community._my_member)
         candidates = self.create_candidates(community, [""] * 5)
@@ -428,7 +430,7 @@ class TestCandidates(DispersyTestFunc):
 
         return community, candidates
 
-    @call_on_mm_thread
+    @call_on_reactor_thread
     def test_tracker_get_introduce_candidate(self, community_create_method=TrackerCommunity.create_community):
         community, candidates = self.test_get_introduce_candidate(community_create_method)
 
@@ -445,7 +447,7 @@ class TestCandidates(DispersyTestFunc):
             got.append(introduce.sock_addr if introduce else None)
         self.assertEquals(expected, got)
 
-    @call_on_mm_thread
+    @call_on_reactor_thread
     def test_introduction_probabilities(self):
         candidates = self.create_candidates(self._community, ["wr", "s"])
         self.set_timestamps(candidates, ["wr", "s"])
@@ -459,7 +461,7 @@ class TestCandidates(DispersyTestFunc):
 
         assert returned_walked_candidate in expected_walked_range
 
-    @call_on_mm_thread
+    @call_on_reactor_thread
     def test_walk_probabilities(self):
         candidates = self.create_candidates(self._community, ["e", "s", "i"])
         self.set_timestamps(candidates, ["e", "s", "i"])
@@ -482,7 +484,7 @@ class TestCandidates(DispersyTestFunc):
         assert returned_stumble_candidate in expected_stumble_range, returned_stumble_candidate
         assert returned_intro_candidate in expected_intro_range, returned_intro_candidate
 
-    @call_on_mm_thread
+    @call_on_reactor_thread
     def test_merge_candidates(self):
         # let's make a list of all possible combinations which should be merged into one candidate
         candidates = []
